Moved all lastfm functions into lastfm.py for organization
