+++ plugins/lastfm.py
@@ -79,3 +79,135 @@ def lastfm(inp, nick='', db=None, bot=None, notice=None):
         db.commit()
 
     return out
+
+
+@hook.command('compare', autohelp=False)
+@hook.command(autohelp=False)
+def compare(inp, nick='', db=None, bot=None, notice=None):
+    """compare [nick] -- Displays the comparison between users
+     from lastfm db."""
+    api_key = bot.config.get("api_keys", {}).get("lastfm")
+    if not api_key:
+        return "error: no api key set"
+
+    qUser = db.execute("select acc from lastfm where nick=lower(?)", (inp,)).fetchone()
+
+    if not qUser:
+        qUser = inp
+    else:
+        qUser = qUser[0]
+
+    nUser = db.execute("select acc from lastfm where nick=lower(?)", (nick,)).fetchone()
+
+    if not nUser:
+        nUser = nick
+    else:
+        nuser = nUser[0]
+
+
+    response = http.get_json(api_url, method="tasteometer.compare",
+                             api_key=api_key, type1="user", type2="user", value1=nuser, value2=qUser, limit=5)
+
+    if 'error' in response:
+        return "Error: {}.".format(response["message"])
+
+    score = round(float(response["comparison"]["result"]["score"]) * 100, 1)
+
+    if score == 0:
+        out = u"You and {} have no artists in common".format(inp)
+    else:
+        out = u"You and {} have {}% in common: artists(".format(qUser, score)
+
+        artists = []
+
+        for item in response["comparison"]["result"]["artists"]["artist"]:
+            artists.append(item["name"])
+
+        lastArtist = artists.pop()
+
+        for artist in artists:
+            out += u"{}, ".format(artist)
+        else:
+            out += u"{})".format(lastArtist)
+
+
+    return out
+
+@hook.command('b', autohelp=False)
+@hook.command(autohelp=False)
+def band(inp, nick='', db=None, bot=None, notice=None):
+    """band [band] -- Displays the band's informations
+     from lastfm db."""
+    api_key = bot.config.get("api_keys", {}).get("lastfm")
+    if not api_key:
+        return "error: no api key set"
+
+    r = http.get_json(api_url, method="artist.getInfo",
+                             api_key=api_key, artist=inp,autocorrect=1,limit=1)
+
+    if 'error' in r:
+        return "Error: {}.".format(r["message"])
+    out="No band named "+ inp
+    tags=[]
+    sims=[]
+
+    if type(r) == dict:
+        artist = r["artist"]
+        if type(artist) ==dict:
+            for tag in artist["tags"]["tag"]:
+                tags.append(tag["name"])
+            for sim in artist["similar"]["artist"]:
+                sims.append(sim["name"])
+            placeformed=" ("+artist["bio"]["placeformed"] +")" if "placeformed" in artist["bio"] else ""
+            out = (artist["name"] + placeformed + " has "+ artist["stats"]["playcount"] + " plays by " + artist["stats"]["listeners"] + " listeners. Tags: " + ", ".join(tags) + ". Similar artists: " + ", ".join(sims) + ". More info on " + artist["url"]).encode("utf-8")
+    return out
+
+@hook.command('top', autohelp=False)
+@hook.command(autohelp=False)
+def top(inp, nick='', db=None, bot=None, notice=None):
+    """top -- Displays the top bands for [nick]
+     from lastfm db."""
+    api_key = bot.config.get("api_keys", {}).get("lastfm")
+    if not api_key:
+        return "error: no api key set"
+
+    if inp:
+        user = db.execute("select acc from lastfm where nick=lower(?)", (inp,)).fetchone()
+    else:
+        user = db.execute("select acc from lastfm where nick=lower(?)", (nick,)).fetchone()
+
+    if not user:
+        if not inp:
+            user = nick
+        else:
+            user = inp
+    else:
+        user = user[0]
+
+    response = http.get_json(api_url, method="user.gettopartists",
+                             api_key=api_key, user=user, period = '7day', limit=5)
+
+    if 'error' in response:
+        return "Error: {}.".format(response["message"])
+
+    topArtists = []
+    lastArtist = ''
+
+    out = u'Top 5 artists this week for {}: ('.format(user)
+
+    if len(response["topartists"]["artist"]) > 1:
+        for artist in response["topartists"]["artist"]:
+            topArtists.append(artist["name"])
+
+        lastArtist = topArtists.pop()
+    else:
+        lastArtist = response["topartists"]["artist"]["name"]
+
+
+    for artist in topArtists:
+        out += u"{}, ".format(artist)
+    else:
+        out += u"{})".format(lastArtist)
+
+    return out
+
